techDebt(buRefactorTraining): Move training toCSV into separate file and out of validator class
